Fix issue where docs of the new version are not published (#277)
Previously, the documentation for the new version had not been published as part of the release. The new docs were only released with the next PR merge after the release was completed. This time gap can be significant. Therefore, releasing the docs has now been made a step in the CD process itself.